+++ idea/src/org/jetbrains/kotlin/idea/findUsages/KotlinNonCodeSearchElementDescriptionProvider.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2010-2016 JetBrains s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jetbrains.kotlin.idea.findUsages
+
+import com.intellij.psi.ElementDescriptionLocation
+import com.intellij.psi.ElementDescriptionProvider
+import com.intellij.psi.PsiElement
+import com.intellij.refactoring.util.NonCodeSearchDescriptionLocation
+import org.jetbrains.kotlin.asJava.namedUnwrappedElement
+import org.jetbrains.kotlin.psi.KtClassOrObject
+
+class KotlinNonCodeSearchElementDescriptionProvider : ElementDescriptionProvider {
+    override fun getElementDescription(element: PsiElement, location: ElementDescriptionLocation): String? {
+        if (location !is NonCodeSearchDescriptionLocation) return null
+        val declaration = element.namedUnwrappedElement as? KtClassOrObject ?: return null
+        return if (location.isNonJava) (declaration.fqName?.asString() ?: declaration.name) else declaration.name
+    }
+}